Rename week 1 workshop to remove focus on Marauroa
The heading for this page makes it appear as though Marauroa is the main thing being learnt about in this
workshop, when in fact it's just an old system we happen to be using to teach other things.  I made some
small edits to change the focus back to the skill being developed, rather than the (incidental) tool we use
for teaching.

(This is also a simple change for me to suggest to start to get used to working with Bookdown and
suggesting changes for your main copy of the text.  Feel free to ignore it if you want.)