Update ind cw 1 instructions for the new cohort
The main change is that students are now required to push their feature branch *before* merging.  This is to bring this coursework more closely into line with the workflow used in the team coursework.

I was not sure where (ref:cautionbox) was defined - presumably in some other file?  (It's used on line 202.)